removed integration own auto updates implementation - lets use HACS instead removed integration codes downloads - codes will be distributed via HACS as well
